refactor(exceptions): clean up diagnostic error imports and remove artifacts
- Remove reverse re-export module diagnostic_errors.py; all imports
  now go through vibe3.exceptions directly
- Fix hardcoded "E_CONFIG_MISSING" string in cli.py, import constant
- Remove merge artifact docs (issue-1898-fix/merge-instructions.md)
- Move test_diagnostics.py from resources/ to utils/ to match source
- Unify all DiagnosticContext/MissingResourceError imports to
  vibe3.exceptions across source and test files

Closes #2050
