+++ src/vibe3/execution/execution_role_policy.py
@@ -12,7 +12,7 @@ from vibe3.config import (
     resolve_effective_agent_options as resolve_backend_effective_agent_options,
 )
 from vibe3.config.convention_resolver import diagnose_profile
-from vibe3.exceptions.diagnostic_errors import DiagnosticContext, MissingResourceError
+from vibe3.exceptions import DiagnosticContext, MissingResourceError
 from vibe3.models import AgentOptions, OrchestraConfig
