+++ src/vibe3/prompts/manifest.py
@@ -11,7 +11,7 @@ import yaml
 from loguru import logger
 
 from vibe3.config.convention_resolver import diagnose_profile
-from vibe3.exceptions.diagnostic_errors import DiagnosticContext, MissingResourceError
+from vibe3.exceptions import DiagnosticContext, MissingResourceError
 from vibe3.prompts.models import (
     LoadedPromptRecipeDefinition,
     PromptRecipeKind,
